requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
factory-boy==3.3.0
Faker==25.9.1
h11==0.14.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.2.2
pytest-django==4.8.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
ruff==0.4.10
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
